+++ src/DotNetty.Common/Utilities/HashedWheelTimer.cs
@@ -15,6 +15,7 @@ namespace DotNetty.Common.Utilities
     using DotNetty.Common.Concurrency;
     using DotNetty.Common.Internal;
     using DotNetty.Common.Internal.Logging;
+    using TaskCompletionSource = DotNetty.Common.Concurrency.TaskCompletionSource;
 
     public sealed class HashedWheelTimer : ITimer
     {