+++ src/DotNetty.Transport/Channels/AbstractChannel.cs
@@ -13,6 +13,7 @@ namespace DotNetty.Transport.Channels
     using DotNetty.Common.Concurrency;
     using DotNetty.Common.Internal.Logging;
     using DotNetty.Common.Utilities;
+    using TaskCompletionSource = DotNetty.Common.Concurrency.TaskCompletionSource;
 
     public abstract class AbstractChannel : DefaultAttributeMap, IChannel
     {