+++ src/DotNetty.Transport/Channels/DefaultChannelPipeline.cs
@@ -16,6 +16,7 @@ namespace DotNetty.Transport.Channels
     using DotNetty.Common.Concurrency;
     using DotNetty.Common.Internal.Logging;
     using DotNetty.Common.Utilities;
+    using ReferenceEqualityComparer = DotNetty.Common.Utilities.ReferenceEqualityComparer;
 
     public class DefaultChannelPipeline : IChannelPipeline
     {