+++ src/DotNetty.Transport/Channels/Sockets/SocketDatagramChannel.cs
@@ -16,6 +16,7 @@ namespace DotNetty.Transport.Channels.Sockets
     using DotNetty.Common.Concurrency;
     using DotNetty.Common.Internal.Logging;
     using DotNetty.Common.Utilities;
+    using TaskCompletionSource = DotNetty.Common.Concurrency.TaskCompletionSource;
 
     public class SocketDatagramChannel : AbstractSocketMessageChannel, IDatagramChannel
     {
@@ -115,7 +116,7 @@ namespace DotNetty.Transport.Channels.Sockets
             operation.SetBuffer(bytes.Array, bytes.Offset, bytes.Count);
 
             bool pending;
-#if NETSTANDARD2_0
+#if NETSTANDARD2_0 || NETCOREAPP3_1_OR_GREATER || NET5_0_OR_GREATER
             pending = this.Socket.ReceiveFromAsync(operation);
 #else
             if (ExecutionContext.IsFlowSuppressed())