+++ test/DotNetty.Codecs.Http.Tests/WebSockets/WebSocketHandshakeHandOverTest.cs
@@ -12,6 +12,7 @@ namespace DotNetty.Codecs.Http.Tests.WebSockets
     using DotNetty.Transport.Channels;
     using DotNetty.Transport.Channels.Embedded;
     using Xunit;
+    using TaskCompletionSource = DotNetty.Common.Concurrency.TaskCompletionSource;
 
     public class WebSocketHandshakeHandOverTest
     {